Validation (#128)
* validator for AddProductsToCartDtoApi has been crteated

* Cart validation has been added

* validation

* Fluent validation has been done

---------

+++ PerfumeStore.API/PerfumeStore.API/Controllers/ProductsController.cs
@@ -1,7 +1,9 @@
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Org.BouncyCastle.Asn1.Ocsp;
 using PerfumeStore.API.Shared.DTO.Request.Product;
+using PerfumeStore.API.Validators;
 using PerfumeStore.Application.Abstractions.Result.Entity;
 using PerfumeStore.Application.Products;
 using PerfumeStore.Application.Shared.DTO.Request;
@@ -17,17 +19,26 @@ namespace PerfumeStore.API.Controllers
     {
         private readonly IProductsService _productService;
         private readonly IMapper _mapper;
+        private readonly IValidationService _validationService;
 
-        public ProductsController(IProductsService productService, IMapper mapper)
+        public ProductsController(IProductsService productService, IMapper mapper, IValidationService validationService)
         {
             _productService = productService;
             _mapper = mapper;
+            _validationService = validationService;
         }
 
         [HttpPost]
         [Authorize(Roles = Roles.Administrator)]
         public async Task<IActionResult> CreateProduct([FromBody] CreateProductDtoApi createProductForm)
         {
+            var validationResult = await _validationService.ValidateAsync(createProductForm);
+
+            if (!validationResult.IsValid)
+            {
+                return BadRequest(validationResult.Errors);
+            }
+
             CreateProductDtoApp createProductDtoApp = _mapper.Map<CreateProductDtoApp>(createProductForm);
 
             EntityResult<ProductResponse> result = await _productService.CreateProductAsync(createProductDtoApp);
@@ -42,6 +53,9 @@ namespace PerfumeStore.API.Controllers
         //[Authorize(Roles.Administrator)]
         public async Task<IActionResult> DeleteProduct(int productId)
         {
+            if (productId <= 0)
+                return BadRequest("Wrong product id");
+
             EntityResult<Product> result = await _productService.DeleteProductAsync(productId);
 
             if (result.IsFailure)
@@ -53,6 +67,9 @@ namespace PerfumeStore.API.Controllers
         [HttpGet("{productId}")]
         public async Task<IActionResult> GetProductById(int productId)
         {
+            if (productId <= 0)
+                return BadRequest("Wrong product id");
+
             EntityResult<ProductResponse> result = await _productService.GetProductByIdAsync(productId);
 
             if (result.IsFailure)
@@ -74,9 +91,15 @@ namespace PerfumeStore.API.Controllers
         }
 
         [HttpPut]
-        //[Authorize(Roles.Administrator)]
         public async Task<IActionResult> UpdateProduct([FromBody] UpdateProductDtoApi updateProductForm)
         {
+            var validationResult = await _validationService.ValidateAsync(updateProductForm);
+
+            if (!validationResult.IsValid)
+            {
+                return BadRequest(validationResult.Errors);
+            }
+
             UpdateProductDtoApp updateProductDtoApp = _mapper.Map<UpdateProductDtoApp>(updateProductForm);
 
             EntityResult<ProductResponse> result = await _productService.UpdateProductAsync(updateProductDtoApp);
